chore: Keil-v5 STM32F103RB -Trevor- FirstProject/Example1 - Simulation
This commit contains all the files needed for the project configuration.
This Files (Project) are tested in microVision V5.28.0.0 .
The Pack with the examples can be downloaded from:
https://booksite.elsevier.com/9780081006290/
The example in the Pack is named:
'Ex 2.1 First Project (Cortex-M_Desingers_Guide)'

As in the book a new Project as to be made right from start with
'New Project' choosing "Blinky.uvprojx". The microcontroller selected
is the 'STM32F103RB' from STMicroelectronics.
In the 'Manage Run-Time Environment' under 'Board Support' the 'MCBSTM32C'
board is chosen. The 'A/D Converter (API)', the 'LED', the 'CMSIS - CORE',
the 'GPIO' and System 'Startup' for STMicroelectronics for STM32F1xx
device series is selected.
Then in the microVision-project window the 'Update Config File
(Ver:0.0.0-> 1.1.2)' is selected for the following files:
RTE_Device.h (Startup)
startup_stm32f10x_md.s (Startup)
system_stm32f10x.c (Startup)
Add the source file 'Blinky.c' with right-click on 'Source Group 1' and
selecting 'Add Existing Files to Group "Source Group 1" '.
Build the Project.

- Options for Target
  -debug -> Use Simulator
  -Parameters: Dialog DLL: DARMRM.DLL - and Parameter: -pSTM32F103RB
  -select Initialization File: Dbg_sim.ini
Build the Project